feat: add zyn-app-map module, optimize system schema, and refactor infrastructure
- Add zyn-app-map module: layers, groups, basemaps, sources, features,
  styles, fields, instances, publishing, versioning, audit, MVT tiles,
  spatial analysis, raster/vector import, tile proxy, GDAL/tippecanoe scripts
- Optimize system schema: rename role_code→code, perm_code→code,
  org_code→code, sort→sort_order, add NOT NULL, fix indexes
- Add file management (sys_file) to zyn-app-system
- Add BizException with typed error codes (404/403/409)
- Centralize cache keys in CacheKeys constants class
- Refactor RedisClient to conditional bean registration
- Fix ApiVersionAutoConfiguration (WebMvcRegistrations instead of
  WebMvcConfigurationSupport)
- Add LoginHelper.init() pattern, SaPermissionImpl safety warning
- Add CI workflow (.github/workflows/ci.yml)
- Add PostgreSQL backup script (deploy/middleware/scripts/backup.sh)
- Add design document (docs/zyn-app-map-design.md)

+++ zyn-api/zyn-api-system/src/main/java/com/zynboot/sys/response/PageRes.java
@@ -16,6 +16,6 @@ public class PageRes<T> {
 
     private List<T> records;
     private long total;
-    private long pageNum;
-    private long pageSize;
+    private int pageNum;
+    private int pageSize;
 }
